+++ modules/apps/wiki/wiki-web/src/main/resources/META-INF/resources/init.jsp
@@ -115,8 +115,8 @@ page import="com.liferay.wiki.model.WikiPageDisplay" %><%@
 page import="com.liferay.wiki.service.WikiNodeServiceUtil" %><%@
 page import="com.liferay.wiki.service.WikiPageLocalServiceUtil" %><%@
 page import="com.liferay.wiki.service.WikiPageServiceUtil" %><%@
-page import="com.liferay.wiki.web.internal.display.context.WikiAdminManagementToolbarDisplayContext" %><%@
-page import="com.liferay.wiki.web.internal.display.context.WikiAdminPagesManagementToolbarDisplayContext" %><%@
+page import="com.liferay.wiki.web.internal.display.context.WikiNodesManagementToolbarDisplayContext" %><%@
+page import="com.liferay.wiki.web.internal.display.context.WikiPagesManagementToolbarDisplayContext" %><%@
 page import="com.liferay.wiki.web.internal.search.NodesChecker" %><%@
 page import="com.liferay.wiki.web.internal.search.PagesChecker" %><%@
 page import="com.liferay.wiki.web.internal.util.WikiPortletUtil" %>
